+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
  - Also allow input of UUIDs instead of names for network, subnet and image based on a UUID regular expression
  - Replace static IP address input with dynamically calculated fixed IP addresses based on subnet cidr and host address index input
+ - Make subnet inputs optional
 
 ## v1.2.1